Safari-609.1.11.1/LayoutTests/fast/sub-pixel/client-width-height-snapping-expected.txt

PASS div.clientWidth is 11
PASS div.clientHeight is 11
PASS successfullyParsed is true

TEST COMPLETE
